Module identity_derive::impls::enums

Functions

derive_diff_enum
derive a Diff type Enum from an incoming InputModel.

impl_debug_enum

impl_diff_enum
derive the Diff trait for incoming Enum in InputModel.

parse_diff
parses data for the derived diff function.

parse_evariants
function that parses and sorts the variants into twp Vec types.

parse_from_into

parse_merge

populate_field_names